### 2025-01-17 #7694 @editorialbot
[PRE REVIEW]: The Population Synthesis Toolkit (PST) Python library
**Submitting author:** @PabloCorcho (<a href="http://orcid.org/0000-0001-6327-7080">Pablo Corcho-Caballero</a>)
**Repository:** https://github.com/paranoya/population-synthesis-toolkit
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.1
