### 2025-01-17 #7694 @warrickball
Hi @PabloCorcho — Thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!
