### 2025-01-29 #7694 @warrickball
I'm afraid this is quite difficult to guess because it depends on how long ongoing reviews take and which editors become available, given their varying expertise. Given the current backlog, however, I don't expect the review to start for at least a few weeks. Sorry!
