### 2025-03-28 #7694 @PabloCorcho
Hi @warrickball, sorry for insisting on this but I'd like to know when do you expect this paper to be reviewed? This package is being used in several papers that are currently being submitted/reviewed (including [early science](https://arxiv.org/abs/2503.15315) with Euclid). I have plans to submit the paper to arxiv as well, but I believe it at least should be under review rather than just submitted. I understand that you're probably overwhelmed, but I'd really appreciate any help on putting this forward.
