### 2025-03-28 #7694 @warrickball
Hi @PabloCorcho, sorry for the wait. We do have quite a heavy backlog at the moment but you're nearing the front of the queue now. A few papers are moving through the publication stages too so I'd hope that an editor will be able to pick this up by the end of April. Note that it will then probably still take a week or two to identify suitable reviewers.

In the meantime, it's quite common for authors to upload the submitted version to arXiv, for which we can generate the appropriate source LaTeX by commenting `@editorialbot generate preprint` in this issue.

If you'd like to pre-empt the sorts of things the reviewers will look for (thus speeding up the review), you can have a look through the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) or the [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html).
