### 2025-03-28 #7694 @warrickball
An implementation quirk: editorialbot commands have to be the first line. (It was years before I learned this myself!)
