### 2025-03-28 #7694 @PabloCorcho
Just a small note, I had to modify the file in order to successfully compile the bibliography. Instead of using `\'` the file uses non-unicode symbols for the tilde (e.g. Mollá instead of `Molla\'`):

`\bibitem[\citeproctext]{ref-molla+09}
Mollá, M., Garcı́a-Vargas, M. L., \& Bressan, A. (2009). {PopStar I:
evolutionary synthesis model description}. \emph{398}(1), 451--470.
\url{https://doi.org/10.1111/j.1365-2966.2009.15160.x}
`
