### 2025-05-12 #7694 @ivastar
@PabloCorcho I will be taking over as an editor. Thank you for your patience!

I will be looking for potential reviewers in the next week or so. If you have recommendations ([from the JOSS reviewers list](https://reviewers.joss.theoj.org/reviewers) or from the broader community) please list the names/handles here or send me an email with your suggestions (momcheva@mpia.de). Your help can greatly expedite the process.
