### 2025-05-15 #8203 @ivastar
@Sujay-Shankar, @jrleja welcome to the review for `PST` and thank you for agreeing to review this JOSS submission. Instructions for generating your review checklists are above and additional information can be found in the linked reviewer guidelines. Don't hesitate to reach out to me via email or by tagging me here if you have any questions or concerns.

The deadline for the first round of reviews is June 15th.
