### 2025-05-30 5f03876
add method for computing time at mass frac
