### 2025-06-06 paranoya/population-synthesis#63 @Sujay-Shankar
Documentation needs some edits
Hi! Hope all is well.
I've read through the nice [documentation](https://population-synthesis-toolkit.readthedocs.io/en/latest/), and while I believe it is mostly complete, there are a few relatively simple quality tweaks I would recommend implementing before I can tick the relevant checkboxes.
I have not yet tested the documentation's code snippets themselves; any issues I find with those will be raised separately.

- [x] The overview section looks good and is very clear on what the code *does*, however it should be clearer in stating the challenges that the software is designed to solve (which are stated in the JOSS paper itself). Additionally, it should be explicit in stating the target audience and intended use case.

### Installation Instructions
- [x] `pyproject.toml` states that the minimum required Python is 3.7, while the documentation states that it is 3.8. This should be made consistent.
- [x] [System Requirements](https://population-synthesis-toolkit.readthedocs.io/en/latest/installation.html#system-requirements) does not list `extinction` and `requests`, which are listed in `requirements.txt`.

### Automated Tests
- [x] I see you have Actions set up and that unit tests are present in the repository, but the documentation should include instructions for users to run the tests themselves to confirm that their installation is working as expected.

### Functionality Documentation
- [x] Many classes in the API reference are missing a complete description of their input parameters, such as [`PopStar`](https://population-synthesis-toolkit.readthedocs.io/en/latest/api.html#pst.SSP.PopStar).

### Quality Edits
- [x] In [Step-by-Step Breakdown](https://population-synthesis-toolkit.readthedocs.io/en/latest/quickstart.html#step-by-step-breakdown), Step 1 describes importing `matplotlib`. However, in [Code Example](https://population-synthesis-toolkit.readthedocs.io/en/latest/quickstart.html#code-example), this line is executed between Steps 6 and 7. The import should be moved up to be with the other imports.
- [x] The expected output plot of [Code Example](https://population-synthesis-toolkit.readthedocs.io/en/latest/quickstart.html#code-example) should be shown in the documentation. In general, wherever there is a plot generated, it should be shown so users can compare.
- [x] In [SSP Features](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#features), code snippets implicitly assume that the user has imported `SSPBase`, created the `ssp_model` variable, and run `import astropy.units as u`. This code snippet should be made explicit; subsequent sections do this correctly.
- [x] In [SSP Features](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#features), Feature 3 uses `u.AA`, while Feature 4 uses `u.angstrom`. These are aliases to the same unit, so I suggest making them consistent.
- [x] Certain Python object names are colored in orange and others are in black. Suggest making the theme consistent.
- [x] In [SSP Features](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#features), the text "example" before code snippets is formatted well. However, in following sections, the formatting of "example" is inconsistent.
- [x] Features are numbered in [SSP Features](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#features) but bulleted in [CEM Features](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#id2).
- [x] The formatting style for expressing function parameters in the API reference varies. For example, under [`SSPBase`](https://population-synthesis-toolkit.readthedocs.io/en/latest/api.html#pst.SSP.SSPBase): `cut_wavelength`, `interpolate_sed`, and `compute_photometry` look different. Suggest making the API reference more consistent in this regard.
- [x] In all math expressions, mathematical functions such as erf() and ln(), as well as text labels such as "today" or "inf", should be formatted in mathrm style, differently from variables.

### Spelling Fixes
- [x] In [Observables](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#observables), "defaul" should be "default" and "downladed" should be "downloaded".
- [x] In [Dust extinction effects](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#dust-extinction-effects), "extinctino" should be "extinction".
- [x] In [Tutorials](https://population-synthesis-toolkit.readthedocs.io/en/latest/tutorials.html), "pratical" should be "practical".
- [x] In [Download SSP models](https://population-synthesis-toolkit.readthedocs.io/en/latest/get_ssp_data.html), "straitforward" should be "straightforward".
- [x] In the [SSPBase API](https://population-synthesis-toolkit.readthedocs.io/en/latest/api.html#pst.SSP.SSPBase), "Paramteres" should be "Parameters".

If I overlooked or misunderstood something, please let me know!

https://github.com/openjournals/joss-reviews/issues/8203
