### 2025-06-07 paranoya/population-synthesis#64 @Sujay-Shankar
JOSS paper suggestions
The paper is well-written, and understandable even to someone like myself that is a non-specialist.
I have a few suggestions for completeness and quality.

### A statement of need
- [x] The statement of need section does clearly state what problems PST solves and its relation to other work, but it needs to be more explicit in stating its target audience. Practitioners in which fields? What expertise levels? What's the use case (as in analysis, visualization, interface, etc.)? That sort of thing.

- [x] There is the sentence fragment: "Compared to alternative approaches in the literature", but this should be elaborated on with an example or two of existing solutions.

The following things should be cited if possible:
- [x] The existing solutions that would be added to address the suggestion above
- [ ] PyKOALA
- [x] BESTA
- [x] CosmoSIS
- [x] SVO Filter Profile Service
- [x] Lick system

### Miscellaneous Edits
- [x] "Cosmosis" should be stylized as "CosmoSIS".
- [x] "Spanish Virtual Observatory Filter Profile Servive" should be "Spanish Virtual Observatory Filter Profile Service"

https://github.com/openjournals/joss-reviews/issues/8203
