+++ paper/paper.bib
@@ -70,3 +70,100 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@INPROCEEDINGS{rodrigo+20,
+       author = {{Rodrigo}, C. and {Solano}, E.},
+        title = "{The SVO Filter Profile Service}",
+    booktitle = {XIV.0 Scientific Meeting (virtual) of the Spanish Astronomical Society},
+         year = 2020,
+        month = jul,
+          eid = {182},
+        pages = {182},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020sea..confE.182R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{worthey+94,
+       author = {{Worthey}, Guy and {Faber}, S.~M. and {Gonzalez}, J. Jesus and {Burstein}, D.},
+        title = "{Old Stellar Populations. V. Absorption Feature Indices for the Complete Lick/IDS Sample of Stars}",
+      journal = {\apjs},
+     keywords = {Absorption Spectra, Sky Surveys (Astronomy), Star Distribution, Abundance, Metallicity, Stellar Atmospheres, Stellar Gravitation, Stellar Temperature, Astronomy, GALAXIES: STELLAR CONTENT, STARS: ABUNDANCES, STARS: ATMOSPHERES, STARS: FUNDAMENTAL PARAMETERS},
+         year = 1994,
+        month = oct,
+       volume = {94},
+        pages = {687},
+          doi = {10.1086/192087},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1994ApJS...94..687W},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{zuntz+15,
+       author = {{Zuntz}, J. and {Paterno}, M. and {Jennings}, E. and {Rudd}, D. and {Manzotti}, A. and {Dodelson}, S. and {Bridle}, S. and {Sehrish}, S. and {Kowalkowski}, J.},
+        title = "{CosmoSIS: Modular cosmological parameter estimation}",
+      journal = {Astronomy and Computing},
+     keywords = {Cosmology:miscellaneous, Methods:data analysis, Methods:statistical, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2015,
+        month = sep,
+       volume = {12},
+        pages = {45-59},
+          doi = {10.1016/j.ascom.2015.05.005},
+archivePrefix = {arXiv},
+       eprint = {1409.3409},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015A&C....12...45Z},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{boquien+19,
+       author = {{Boquien}, M. and {Burgarella}, D. and {Roehlly}, Y. and {Buat}, V. and {Ciesla}, L. and {Corre}, D. and {Inoue}, A.~K. and {Salas}, H.},
+        title = "{CIGALE: a python Code Investigating GALaxy Emission}",
+      journal = {\aap},
[…]
