+++ paper/paper.bib
@@ -150,7 +150,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{cc+25,
-       author = {{Euclid Collaboration} and {Corcho-Caballero}, P. and {Ascasibar}, Y. and {Verdoes Kleijn}, G. and {Lovell}, C.~C. and {De Lucia}, G. and {Cleland}, C. and {Fontanot}, F. and {Tortora}, C. and {Koopmans}, L.~V.~E. and {Moutard}, T. and {Laigle}, C. and {Nersesian}, A. and {Shankar}, F. and {Aghanim}, N. and {Altieri}, B. and {Amara}, A. and {Andreon}, S. and {Aussel}, H. and {Baccigalupi}, C. and {Baldi}, M. and {Balestra}, A. and {Bardelli}, S. and {Battaglia}, P. and {Biviano}, A. and {Bonchi}, A. and {Bonino}, D. and {Branchini}, E. and {Brescia}, M. and {Brinchmann}, J. and {Ca{\~n}as-Herrera}, G. and {Capobianco}, V. and {Carbone}, C. and {Carretero}, J. and {Casas}, S. and {Castander}, F.~J. and {Castellano}, M. and {Castignani}, G. and {Cavuoti}, S. and {Chambers}, K.~C. and {Cimatti}, A. and {Colodro-Conde}, C. and {Congedo}, G. and {Conselice}, C.~J. and {Conversi}, L. and {Copin}, Y. and {Costille}, A. and {Courbin}, F. and {Courtois}, H.~M. and {Cropper}, M. and {Da Silva}, A. and {Degaudenzi}, H. and {Di Giorgio}, A.~M. and {Dolding}, C. and {Dole}, H. and {Dubath}, F. and {Dupac}, X. and {Ealet}, A. and {Escoffier}, S. and {Farina}, M. and {Farinelli}, R. and {Faustini}, F. and {Ferriol}, S. and {Finelli}, F. and {Fotopoulou}, S. and {Frailis}, M. and {Franceschi}, E. and {Fumana}, M. and {Galeotta}, S. and {George}, K. and {Gillis}, B. and {Giocoli}, C. and {Gracia-Carpio}, J. and {Granett}, B.~R. and {Grazian}, A. and {Grupp}, F. and {Guzzo}, L. and {Gwyn}, S. and {Haugan}, S.~V.~H. and {Holmes}, W. and {Hook}, I.~M. and {Hormuth}, F. and {Hornstrup}, A. and {Hudelot}, P. and {Jahnke}, K. and {Jhabvala}, M. and {Keih{\"a}nen}, E. and {Kermiche}, S. and {Kiessling}, A. and {Kubik}, B. and {Kuijken}, K. and {K{\"u}mmel}, M. and {Kunz}, M. and {Kurki-Suonio}, H. and {Le Boulc'h}, Q. and {Le Brun}, A.~M.~C. and {Le Mignant}, D. and {Ligori}, S. and {Lilje}, P.~B. and {Lindholm}, V. and {Lloro}, I. and {Mainetti}, G. and {Maino}, D. and {Maiorano}, E. and {Mansutti}, O. and {Marcin}, S. and {Marggraf}, O. and {Martinelli}, M. and {Martinet}, N. and {Marulli}, F. and {Massey}, R. and {Maurogordato}, S. and {Medinaceli}, E. and {Mei}, S. and {Melchior}, M. and {Mellier}, Y. and {Meneghetti}, M. and {Merlin}, E. and {Meylan}, G. and {Mora}, A. and {Moresco}, M. and {Moscardini}, L. and {Nakajima}, R. and {Neissner}, C. and {Niemi}, S. -M. and {Nightin
[…]
