### 2025-06-09 3f88909
full documentation of SSP models
