+++ src/pst/SSP.py
@@ -319,8 +319,60 @@ class SSPBase(object):
 
 
 class PopStar(SSPBase):
-    """PopStar SSP models (Mollá+09)."""
+    """
+    PopStar SSP models (Mollá et al. 2009).
 
+    This class represents the PopStar evolutionary synthesis models for simple
+    stellar populations as presented in Mollá et al. (2009). It provides
+    access to spectral energy distributions (SEDs) across a range of metallicities
+    and ages, optionally including nebular emission.
+
+    Parameters
+    ----------
+    IMF : str
+        Initial Mass Function name. Must be one of the supported IMFs listed below:
+
+        - 'sal' : Salpeter (1955)
+        - 'fer' : Ferrini, Penco & Palla (1990)
+        - 'kro' : Kroupa (2002)
+        - 'cha' : Chabrier (2003)
+
+        This keyword is used to select the corresponding PopStar model files.
+    nebular : bool, optional
+        If True, include nebular emission in the SEDs. Default is False, meaning
+        only the stellar continuum is considered.
+    path : str or None, optional
+        Filesystem path to the PopStar model data. If None (default), the package
+        default path plus 'PopStar' subdirectory is used.
+    verbose : bool, optional
+        If True (default), print informational messages during initialization.
+
+    Example
+    -------
+    Initialize with the desired IMF and nebular option:
+
+    >>> from pst.SSP import PopStar
+    >>> help(PopStar) # Retrieve the class documentation
+    >>> popstar = PopStar(IMF="cha", nebular=True) # Initialise the model
+    >>> print(popstar.wavelength)
+    <Quantity [91. , 94. , 96. , ..., 1200000., 1400000., 1600000.] Angstrom>
+    >>> oldest_sed = popstar.L_lambda[-1, -1] # Get the SED of the oldest SSP with highest metallicity
+
+    Notes
+    -----
+    - The PopStar models cover metallicities from 0.0001 to 0.05 and ages spanning from
+      10^5 to ~10^10 years.
+    - Nebular emission is included only if `nebular=True` is set.
+    - Model files are expected to be in compressed FITS format with specific naming
+      conventions as per Mollá et al. 2009.
+
+    For more details see the :class:`SSPBase` base class documentation.
+
+    References
+    ----------
+    Mollá, M., García-Vargas, M. L., & Bressan, A. (2009). PopStar I: evolutionary synthesis model description. `MNRAS, 398(1), 451-470 <https://academic.oup.com/mnras/article/398/1/451/1099698>`_.
+    `Fractal data repository <https://www.fractal-es.com/PopStar/>`_
+    """
     def __init__(self, IMF, nebular=False, path=None, verbose=True):
         if path is None:
             self.path = os.path.join(self.default_path, 'PopStar')
@@ -377,8 +429,57 @@ class PopStar(SSPBase):
 
 
 class PyPopStar(SSPBase):
-    """PyPopStar SSP models (Millán-Irigoyen+21)."""
+    """PyPopStar SSP models (Millán-Irigoyen et al. 2021).
 
+    This class represents the PyPopStar simple stellar population (SSP) models
+    as described in Millán-Irigoyen et al. (2021). It provides spectral energy
+    distributions (SEDs) for a range of metallicities and ages, optionally
+    including nebular emission.
+
+    Parameters
+    ----------
+    IMF : str
+        Initial Mass Function identifier. Must be one of the supported IMFs:
+
+        - 'KRO' : Kroupa (2002)
+        - 'SAL' : Salpeter (1955)
+        - 'CHA' : Chabrier (2003)
+        This keyword selects the corresponding model subdirectory and file set.
+    nebular : bool, optional
[…]
