### 2025-06-09 bc3922d
implement rev corrections
