### 2025-06-09 bbaaab3
homogeneize documentation
