### 2025-06-09 7b7e734
add additional requirements
