### 2025-06-09 paranoya/population-synthesis-toolkit#63 @PabloCorcho
Dear @Sujay-Shankar,

Many thanks for reviewing the paper and providing comments.

The suggested changes pertaining documentation are implemented via #66. Let me provide some notes here to facilitate the review. You may tick off the comments listed above when you consider they've been implemented in the PR.

> ### Statement of Need
>
>     * [ ]  The overview section looks good and is very clear on what the code _does_, however it should be clearer in stating the challenges that the software is designed to solve (which are stated in the JOSS paper itself). Additionally, it should be explicit in stating the target audience and intended use case.
>

I have included a few extra sentences [here](https://github.com/paranoya/population-synthesis-toolkit/pull/66/files#diff-4eec6cec5f6fab1548b85433ea8ca81315ae165db4b7f84019f287df9015699f).

>
> ### Installation Instructions
>
>     * [ ]  `pyproject.toml` states that the minimum required Python is 3.7, while the documentation states that it is 3.8. This should be made consistent.

Set to Python 3.8

>     * [ ]  [System Requirements](https://population-synthesis-toolkit.readthedocs.io/en/latest/installation.html#system-requirements) does not list `extinction` and `requests`, which are listed in `requirements.txt`.
>
>

Included.

> ### Automated Tests
>
>     * [ ]  I see you have Actions set up and that unit tests are present in the repository, but the documentation should include instructions for users to run the tests themselves to confirm that their installation is working as expected.
>

I have included some [instructions](https://github.com/paranoya/population-synthesis-toolkit/pull/66/files#diff-e285c1b9901cd202e89d9f37e3249a6fee524a65428895e1a45fd062cc0f9aca), and a bash script for running all the tests, which avoids having to require installing e.g., `pytest`.

>
> ### Functionality Documentation
>
>     * [ ]  Many classes in the API reference are missing a complete description of their input parameters, such as [`PopStar`](https://population-synthesis-toolkit.readthedocs.io/en/latest/api.html#pst.SSP.PopStar).
>

I have significantly expanded the [documentation](https://github.com/paranoya/population-synthesis-toolkit/pull/66/files#diff-bf1d2373b25c799725266aad0611d00d3d3fb25cb83d4314ed0e66523071ab76) of all SSP models currently used, providing a detailed description about the model origin and its main properties.

>
> ### Quality Edits
>
>     * [ ]  In [Step-by-Step Breakdown](https://population-synthesis-toolkit.readthedocs.io/en/latest/quickstart.html#step-by-step-breakdown), Step 1 describes importing `matplotlib`. However, in [Code Example](https://population-synthesis-toolkit.readthedocs.io/en/latest/quickstart.html#code-example), this line is executed between Steps 6 and 7. The import should be moved up to be with the other imports.
>     * [ ]  The expected output plot of [Code Example](https://population-synthesis-toolkit.readthedocs.io/en/latest/quickstart.html#code-example) should be shown in the documentation. In general, wherever there is a plot generated, it should be shown so users can compare.

[Both done.](https://github.com/paranoya/population-synthesis-toolkit/pull/66/files#diff-a106101eaf2135a4c111a1f2d184cb7bc69ab453c0a7353951e307e7a088cf94)

>     * [ ]  In [SSP Features](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#features), code snippets implicitly assume that the user has imported `SSPBase`, created the `ssp_model` variable, and run `import astropy.units as u`. This code snippet should be made explicit; subsequent sections do this correctly.
>     * [ ]  In [SSP Features](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#features), Feature 3 uses `u.AA`, while Feature 4 uses `u.angstrom`. These are aliases to the same unit, so I suggest making them consistent.

[Both done.](https://github.com/paranoya/population-synthesis-toolkit/pull/66/files#diff-316442bb11c260938e4dfe087b6e5b1e2301f7440b527a41157ba330ae5b5be4)

>     * [ ]  Certain Python object names are colored in orange and others are in black. Suggest making the theme consistent.

Now, only class methods appear in orange, while the class references to the API documentation are shown in black.

>     * [ ]  In [SSP Features](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#features), the text "example" before code snippets is formatted well. However, in following sections, the formatting of "example" is inconsistent.

Fixed.

>     * [ ]  Features are numbered in [SSP Features](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#features) but bulleted in [CEM Features](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#id2).

Everything is bullet-listed now.

>     * [ ]  The formatting style for expressing function parameters in the API reference varies. For example, under [`SSPBase`](https://population-synthesis-toolkit.readthedocs.io/en/latest/api.html#pst.SSP.SSPBase): `cut_wavelength`, `interpolate_sed`, and `compute_photometry` look different. Suggest making the API reference more consistent in this regard.

I have homogenized the methods documentation to the standard followed throughout the whole repository.

>     * [ ]  In all math expressions, mathematical functions such as erf() and ln(), as well as text labels such as "today" or "inf", should be formatted in mathrm style, differently from variables.
>

All functions and subscripts appear now in rm style.

>
> ### Spelling Fixes
>
>     * [ ]  In [Observables](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#observables), "defaul" should be "default" and "downladed" should be "downloaded".[ ]  In [Dust extinction effects](https://population-synthesis-toolkit.readthedocs.io/en/latest/user_guide.html#dust-extinction-effects), "extinctino" should be "extinction".[ ]  In [Tutorials](https://population-synthesis-toolkit.readthedocs.io/en/latest/tutorials.html), "pratical" should be "practical".[ ]  In [Download SSP models](https://population-synthesis-toolkit.readthedocs.io/en/latest/get_ssp_data.html), "straitforward" should be "straightforward".[ ]  In the [SSPBase API](https://population-synthesis-toolkit.readthedocs.io/en/latest/api.html#pst.SSP.SSPBase), "Paramteres" should be "Parameters".
>

Corrected.

Let me know if there's any additional change I should implement.
