### 2025-06-13 paranoya/population-synthesis-toolkit#67 @Sujay-Shankar
Unit testing need some small tweaks
I've tested PST installation and unit tests on Python 3.8 through 3.13 with both `mamba` via WSL and `uv` via macOS.
Installation proceeds smoothly, and all unit tests pass successfully, however there are a few small tweaks that are needed before I can tick the "Installation" checkbox.

- [ ] The invocation of `run_tests.sh` needs to be changed to `bash run_tests.sh`. Otherwise you get a "permission denied" error.
- [ ] On Python 3.12 or above, you get a `SyntaxWarning` when running unit tests for invalid escape sequences. This is caused by docstrings with math expressions having backslashes. These docstrings need to be raw strings so that the interpreter doesn't look for escape sequences.

https://github.com/openjournals/joss-reviews/issues/8203
