### 2025-06-13 paranoya/population-synthesis#68 @jrleja
JOSS Paper Review
The paper is succinct and clear. I tabulate below a few points to address so that I can check off the appropriate review boxes.

**A statement of need + State of the field + References**
Missing is a substantive comparison of the features and purpose Population Synthesis Toolkit to other tools in the field. Here is a short, incomplete list of codes that come to mind:

- DSPS: https://github.com/ArgonneCPAC/dsps
- python-fsps: https://github.com/dfm/python-fsps
- ProSpect: https://github.com/asgr/ProSpect
- SPISEA: https://github.com/astropy/SPISEA
- binary_c-python: https://gitlab.surrey.ac.uk/ri0005/binary_c-python

Pointing to similar codebases and clarifying how the Population Synthesis Toolkit fits in will be very helpful.

Also, a few suggestions on grammar and jargon:
- dust extinction --> extinction and attenuation. Extinction refers to line-of-sight dust extinction, whereas attenuation is the net effect of extinction and scattering for a resolved system (typically a galaxy). Your code has both, through the `extinction` package, e.g. the Calzetti+00 attenuation curve.
- equivalent widths of **(emission/absorption?)** lines. (I assume absorption).
- I would suggest considering a name change to the package for disambiguation purposes, e.g. "Stellar Population Synthesis Toolkit". This is not at all a requirement for JOSS publication! Here's the reason: "population synthesis" in astronomy refers to a variety of techniques to create populations (e.g. neutron stars, SMBHs, stellar populations, ...), and without a modifier out front, generally refers to _binary_ population synthesis (BPASS happens to do both, and they have to use even more precise language to disambiguate). So this would be helpful for disambiguation and for ensuring you are connected to the correct audience.

https://github.com/openjournals/joss-reviews/issues/8203
