### 2025-06-13 paranoya/population-synthesis#69 @jrleja
Documentation suggestions
Hi,

Everything looks like it's in pretty great shape! Thanks for making this useful package. A few suggestions below to maximize the value of the documentation:

- in quickstart guide code example: should be ssp_model.metallicities.value. similarly in the second to last block of analytical.ipynb. (This throws an error on my machine otherwise).
- good practice to show the plot and/or outputs the user should be receiving, in jupyter notebooks and readthedocs code examples, so the user can verify that this is being correctly produced on their machine.
- It would be very helpful if you added a demonstration of using the Lick indexes.

Joel
