### 2025-06-13 #8203 @jrleja
I've completed my first round of review. Many of my initial comments have been covered already by the pull request from Sanjay's review: https://github.com/paranoya/population-synthesis-toolkit/pull/66. Thank you! I cannot easily tell if the API changes are sufficient until this pull request is approved though at a quick glance, they seem excellent. I'll return for this.

I think the software paper would benefit from a few more sentences in the comparison to the state of the field and references, and have suggested specific additions: https://github.com/paranoya/population-synthesis-toolkit/issues/68

I have some suggestions and bugs in the documentation and code examples: https://github.com/paranoya/population-synthesis-toolkit/issues/69. But these are small/nonessential and do not prevent me from checking any review boxes.
