### 2025-06-13 #8203 @Sujay-Shankar
I think I've completed my first round as well! Most of the tweaks should be fairly straightforward I think, the library was already in pretty good shape to begin with. A lot of my comments have to do with the documentation, so I'm eagerly awaiting the new version to go up online to see that next level of polish. With those changes addressed, I'd tick the remaining checkboxes.
And thanks to Pablo for your detailed responses to these issues! Really helps us stay organized and see all the changes have been made.
