### 2025-06-15 8f138ab
fix equation
