+++ src/pst/models.py
@@ -16,6 +16,13 @@ class MassPropMetallicityMixin:
     
     .. math::
         Z(t) = Z_{\rm today} \cdot \left(\frac{M_{\star}(t)}{M_{\star}(\rm today)}\right)^\alpha
+        :label: powlaw-metallicity
+
+    where :math:`Z_{\rm today}` is the ISM metallicity at present,
+    :math:`M_{\star}(t)` is the stellar mass formed at time :math:`t`, and
+    :math:`M_{\star}(\rm today)` is the stellar mass at present.
+    The exponent :math:`\alpha` regulates the chemical enrichment rate with stellar mass.
+
     """
     @property
     def ism_metallicity_today(self):
@@ -36,6 +43,18 @@ class MassPropMetallicityMixin:
         self._alpha_powerlaw = value
 
     def ism_metallicity(self, times):
+        """Evaluate the ISM metallicity at a given set of times, see eq. :eq:`powlaw-metallicity`.
+
+        Parameters
+        ----------
+        times : :class:`astropy.units.Quantity`
+            Cosmic times at which the metallicity will be evaluated.
+
+        Returns
+        -------
+        z_t : :class:`astropy.units.Quantity`
+            Vector with the ISM metallicity at each input time.
+        """
         m = self.stellar_mass_formed(times)
         return self.ism_metallicity_today * np.power(m / self.mass_today, self.alpha_powerlaw)
 
@@ -65,17 +84,18 @@ class ChemicalEvolutionModel(ABC):
         pass
 
     @abstractmethod
-    def stellar_mass_formed(self, time):
+    def stellar_mass_formed(self, time) -> u.Quantity:
         """Total stellar mass formed at a given time."""
         return
 
     @abstractmethod
-    def ism_metallicity(self, time):
+    def ism_metallicity(self, time) -> u.Quantity:
         """ISM metals mass fraction at a given time."""
         return
 
     @u.quantity_input
-    def interpolate_ssp_masses(self, ssp: SSPBase, t_obs: u.Gyr, oversample_factor=10):
+    def interpolate_ssp_masses(self, ssp: SSPBase, t_obs: u.Gyr,
+                               oversample_factor=10) -> u.Quantity:
         """
         Interpolate the star formation history to compute the SSP stellar masses.
 
@@ -85,7 +105,7 @@ class ChemicalEvolutionModel(ABC):
 
         Parameters
         ----------
-        SSP : pst.SSP.SSPBase
+        SSP : :class:`pst.SSP.SSPBase`
             The Simple Stellar Population (SSP) model used for synthesizing the SED.
         t_obs : astropy.Quantity
             The cosmic time at which the galaxy is observed. Only SSPs with ages
@@ -122,7 +142,7 @@ class ChemicalEvolutionModel(ABC):
                                masses=bin_mass)
 
     def compute_SED(self, ssp : SSPBase, t_obs : u.Quantity,
-                    allow_negative=False):
+                    allow_negative=False) -> u.Quantity:
         """
         Compute the Spectral Energy Distribution (SED) resulting from the SFH.
 
@@ -157,14 +177,12 @@ class ChemicalEvolutionModel(ABC):
                     axis=(0))
         return sed
 
-    def compute_photometry(self, ssp, t_obs, photometry=None):
+    def compute_photometry(self, ssp, t_obs, photometry=None) -> u.Quantity:
         """
-        Compute the syntehtic photometry using a SSP at a given time.
+        Compute the synthetic photometry using a SSP at a given time.
 
         This method computes the photometric fluxes associated to
         the SFH by synthesizing the fluxes of the input SSP model.
[…]
