### 2025-06-15 241c275
add numpy import
