### 2025-06-15 b0d7dc9
add script for making pltos
