### 2025-06-15 paranoya/population-synthesis-toolkit#70 @PabloCorcho
>     1. When initializing an `ExponentialDelayedCEM` following the user guide, the following line:
>
>
> ```
> cem_model = ExponentialDelayedCEM(stellar_mass_inf=1.0, tau=3.0, metallicity=0.02)
> ```
>
> throws a `KeyError`, as it is missing the `today` keyword argument. It is also missing `mass_today` and `ism_metallicity_today` arguments. Recommend adding these three arguments to the example.
>

The code example was outdated. This is now fixed in #66.

>     2. Every single time a `matplotlib` plot is generated in example code, it should A. have `plt.show()` called after plot creation, and B. Have corresponding images provided so users can compare. This was done successfully in the Quickstart guide.

Both features are now included.

>     3. There are a lot of variables missing definitions as well as missing imports in code snippets (more than I thought previously). To address this in depth, I would recommend you load the entirety of the User Guide code into say, a Jupyter notebook, and run the code snippets in the User Guide one after the other. Ideally, these should run flawlessly without any errors or warnings. Assume that each page of the documentation runs on its own Python kernel; code in the User Guide should not presume the definition of variables from the Quickstart Guide (a different page of the documentation).

The new version of the page code examples is self-consistent.

>     4. Perhaps some example code snippets and plots should be included for the `dust` module to showcase its functionality and for completeness.

I provided a link to the notebook tutorial with a “real” example of using this feature.
