+++ paper/paper.bib
@@ -166,4 +166,11 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-
+@misc{pykoala,
+  author = {{Corcho-Caballero}, P. and {Ascasibar}, Y. and {L\'opez-S\'anchez}, A. R. and {Gonz\'alez-Bol\'ivar}, M., and {Lorente}, N., and {Tocknell}, J. and {Jim\'enez-Ibarra}, F. and {Daluwathumullagamage}, P. J. and {Quattropani}, G. and {Owers}, M},
+  title = {PyKOALA},
+  year = {2025},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/pykoala}
+}
