### 2025-06-17 2b37d37
add bash
