### 2025-06-17 b6316f2
convert docstrings into raw docstrings
