+++ src/pst/observables.py
@@ -212,8 +212,9 @@ class Filter(object):
         Description
         -----------
         The effective wavelength is computed as
-        .. :math:
-            eff_wl = \frac{\int{R(\lambda) \cdot \lambda d\lambda}}{\int{R(\lambda) d\lambda}}
+
+        .. math::
+            \lambda_{\rm eff} = \frac{\int{R(\lambda) \cdot \lambda d\lambda}}{\int{R(\lambda) d\lambda}}
         
         Returns
         -------
@@ -228,8 +229,9 @@ class Filter(object):
         Description
         -----------
         The effective bandwith is computed as
-        .. :math:
-            eff_bw = \sqrt{8\log(2)} \left(\frac{\int{R(\lambda) \cdot \lambda^2 d\lambda}}{\int{R(\lambda) d\lambda}} - eff_wl\right)^{1/2}
+
+        .. math::
+            \lambda_{\rm BW} = \sqrt{8\log(2)} \left(\frac{\int{R(\lambda) \cdot \lambda^2 d\lambda}}{\int{R(\lambda) d\lambda}} - \lambda_{\rm eff}\right)^{1/2}
 
         Returns
         -------
@@ -250,8 +252,9 @@ class Filter(object):
         Description
         -----------
         The effective transmission is computed as
-        .. :math:
-            \frac{\int{R(\lambda)^2 d\lambda}}{\int{R(\lambda) d\lambda}}
+
+        .. math::
+            R_{\rm eff} = \frac{\int{R(\lambda)^2 d\lambda}}{\int{R(\lambda) d\lambda}}
 
         Returns
         -------
@@ -529,16 +532,35 @@ class TopHatFilter(Filter):
 
 
 class EquivalentWidth(object):
-    """Equivalent width of an spectral region.
+    r"""Equivalent width of an spectral region.
     
-    Attributes
-    ----------
-    left_wl_range : :class:`np.ndarray` or :class:``astropy.units.Quantity``
-        Spectral range defining the left pseudocontinuum window.
-    right_wl_range : :class:`np.ndarray` or :class:``astropy.units.Quantity``
-        Spectral range defining the right pseudocontinuum window.
-    central_wl_range : :class:`np.ndarray` or :class:``astropy.units.Quantity``
-        Spectral range defining the equivalent width window.
+    Description
+    -----------
+    Given a stellar spectra spectra :math:`F_\lambda`, the equivalent width is
+    defined as the area of the spectral line, defined over a given spectral
+    region (:math:`\lambda_{\rm C,\,min}, \lambda_{\rm C,\,max}`), divided by
+    the average flux of the continuum in a given spectral region.
+    It is computed as: 
+
+    .. math::
+        EW = \int_{\lambda_{C,\,\rm min}}^{\lambda_{\rm C,\,max}} \left(1 - \frac{F_\lambda(\lambda)}{F_{\rm cont}(\lambda)}\right) d\lambda
+    where:
+
+    .. math::
+        F_{cont}(\lambda) = \frac{F_{\rm B} \lambda_{\rm R} - F_{\rm R} \lambda_{\rm B}}{\lambda_{\rm R} - \lambda_{\rm B}} + \lambda \frac{F_{\rm R} - F_{\rm B}}{\lambda_{\rm R} - \lambda_{\rm B}}
+        
+    and :math:`F_{\rm B}` and :math:`F_{\rm R}` are the average flux in the left
+    and right spectral windows, respectively, defined as:
+
+    .. math::
[…]
