### 2025-06-18 081ffcc
expand ew documentation
