### 2025-06-18 a9abb7f
expand ew documentation
