### 2025-06-18 9b7bb78
include references to other codes
