+++ paper/paper.md
@@ -44,9 +44,9 @@ PST is a Python library that offers a comprehensive and flexible framework for s
 
 # State of the field
 
-A number of software packages have been developed to support stellar population synthesis and modeling of galaxy spectral energy distributions. Tools such as [binary_c-python](https://gitlab.surrey.ac.uk/ri0005/binary_c-python) and [SPISEA](https://github.com/astropy/SPISEA) [@hosek+20] are primarily designed for generating and analyzing simple stellar populations, often with a focus on individual stars, binaries, or star clusters. Meanwhile, libraries such as [python-FSPS](https://github.com/dfm/python-fsps), a Python interface to the Flexible Stellar Population Synthesis (FSPS) code [@conroy+09], and the more recent [DSPS](https://github.com/ArgonneCPAC/dsps) [@hearin+23], implemented using JAX for efficient gradient computation and forward modeling, provide extensive SSP modeling capabilities.
+A number of software packages have been developed to support stellar population synthesis and modeling of galaxy spectral energy distributions. Tools such as [binary_c-python](https://gitlab.surrey.ac.uk/ri0005/binary_c-python) and [SPISEA](https://github.com/astropy/SPISEA) [@hosek+20] are primarily designed for generating and analyzing simple stellar populations, often with a focus on individual stars, binaries, or star clusters. Meanwhile, libraries such as [python-FSPS](https://github.com/dfm/python-fsps), a Python interface to the Flexible Stellar Population Synthesis (FSPS) code [@conroy+09], and the more recent [DSPS](https://github.com/ArgonneCPAC/dsps) [@hearin+23], implemented using JAX for efficient gradient computation and forward modeling, provide extensive modeling capabilities, although they are sometimes limited to a specific set of SSP models or isochrones.
 
-Other packages put a stronger emphasis on fitting observed data to derive galaxy properties. These include Bayesian frameworks like [CIGALE](https://cigale.lam.fr/) [@boquien+19], [ProSpect](https://github.com/asgr/ProSpect) [@robotham+20] or [Prospector](https://prospect.readthedocs.io/en/v1.0.0/) [@johnson+21], which infer star formation histories and other physical parameters using spectro-photometric data. Alternative frequentist tools such as [Starlight](http://www.starlight.ufsc.br/) (@cid-fernandes+05), [PpXF](https://pypi.org/project/ppxf/) [capellari+04], or [Pipe3D](https://gitlab.com/pipe3d/pyPipe3D) [@sanchez+16], are commonly used to extract stellar kinematics and stellar population parameters from observed galaxy spectra, often in the context of integral field spectroscopy.
+Other packages put a stronger emphasis on fitting observed data to derive galaxy properties. These include Bayesian frameworks like [CIGALE](https://cigale.lam.fr/) [@boquien+19], [ProSpect](https://github.com/asgr/ProSpect) [@robotham+20] or [Prospector](https://prospect.readthedocs.io/en/v1.0.0/) [@johnson+21], which infer star formation histories and other physical parameters using spectro-photometric data. Alternative frequentist tools such as [Starlight](http://www.starlight.ufsc.br/) [@cid-fernandes+05], [PpXF](https://pypi.org/project/ppxf/) [@capellari+04], or [Pipe3D](https://gitlab.com/pipe3d/pyPipe3D) [@sanchez+16], are commonly used to extract stellar kinematics and stellar population parameters from observed galaxy spectra, often in the context of integral field spectroscopy.
 
 # Statement of need
 
