### 2025-06-19 59903c2
remove rogue code
