### 2025-06-19 d21fc84
change equation symbol
