### 2025-06-19 fc82e9e
clarify EW signs
