### 2025-06-19 paranoya/population-synthesis-toolkit#68 @PabloCorcho
Hi @jrleja

Thank you very much for your comments. I am implementing these suggestions through #73.

> **A statement of need + State of the field + References** Missing is a substantive comparison of the features and purpose Population Synthesis Toolkit to other tools in the field. Here is a short, incomplete list of codes that come to mind:
>
>     * DSPS: https://github.com/ArgonneCPAC/dsps
>
>     * python-fsps: https://github.com/dfm/python-fsps
>
>     * ProSpect: https://github.com/asgr/ProSpect
>
>     * SPISEA: https://github.com/astropy/SPISEA
>
>     * binary_c-python: https://gitlab.surrey.ac.uk/ri0005/binary_c-python
>
>
> Pointing to similar codebases and clarifying how the Population Synthesis Toolkit fits in will be very helpful.
>

Yes, you are absolutely right. I included an additional section (State of the field) where I discuss other available tools, providing links to their repositories/online documentation and references. I have divided the software into two categories: code devoted for modelling SEDs of SSP/CSP, and code used for fitting observational data. Let me know if you have any further recommendation.

>     * dust extinction --> extinction and attenuation.

I have explicitly mentioned this on the new version of the paper.

>     * equivalent widths of **(emission/absorption?)** lines. (I assume absorption).
>

It works with both, the only difference being the sign of the EW (i.e. $\rm EW_{abs}> 0$, $\rm EW_{emi}< 0$). I have noted that in the *Features* section and the docstrings of `EquivalentWidth`.

>     * I would suggest considering a name change to the package for disambiguation purposes, e.g. "Stellar Population Synthesis Toolkit". This is not at all a requirement for JOSS publication! Here's the reason: "population synthesis" in astronomy refers to a variety of techniques to create populations (e.g. neutron stars, SMBHs, stellar populations, ...), and without a modifier out front, generally refers to _binary_ population synthesis (BPASS happens to do both, and they have to use even more precise language to disambiguate). So this would be helpful for disambiguation and for ensuring you are connected to the correct audience.
>

Thanks for the suggestion. While there is potential ambiguity, our plan is to progressively expand the library’s capabilities beyond stellar populations to include additional components such as gas, dust, and cosmic rays. We think Population, rather than Stellar population will ultimately better reflect the library’s intended range.

To mitigate potential confusion, both the repository README and the home documentation page explicitly mention that the current scope of the library is stellar population synthesis.
