### 2025-06-19 #8203 @PabloCorcho
Hi @jrleja @Sujay-Shankar,

Once again, many thanks for your feedback on both the paper and the codebase.

To summarize, the paper-related changes have been implemented via:

- [PR65](https://github.com/paranoya/population-synthesis-toolkit/pull/65)
- [PR65](https://github.com/paranoya/population-synthesis-toolkit/pull/73)

You can access the latest version of the paper draft (paper.zip) here:
https://github.com/paranoya/population-synthesis-toolkit/actions/runs/15752159181/artifacts/3361041096

Updates to the source code (including improvements to the online documentation, and docstrings) were made through:

- [PR72](https://github.com/paranoya/population-synthesis-toolkit/pull/72)
- [PR72](https://github.com/paranoya/population-synthesis-toolkit/pull/66)

Further implementation details can be found in the corresponding GitHub issues you opened.

Let me know if there are additional modifications required.

Best wishes,
Pablo
