### 2025-07-03 #8203 @PabloCorcho
> :point_right::page_facing_up: [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08203/joss.08203/10.21105.joss.08203.pdf) :page_facing_up: [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.08203/joss.08203/10.21105.joss.08203.pdf) :page_facing_up: :point_left:

This PDF version doesn't include the changes that were implemented on the PR mentioned above. I have just merged both of them so you can generate the latest draft version.
