### 2025-07-03 #8203 @rhdtownsend
Thanks — indeed I’ve not been able to find time to contribute, apologies. Best you proceed without me.

> On Jul 3, 2025, at 4:46 AM, Ivelina Momcheva ***@***.***> wrote:
>
> ivastar left a comment (openjournals/joss-reviews#8203)
> @rhdtownsend pinging you about this review if you are still interested and especially if there is anything you think the other reviewers have not yet covered. But it's ok if you don't have time, let me know.
