### 2025-07-11 #8203 @ivastar
@jrleja @Sujay-Shankar thank you for the responses!

@PabloCorcho I will do an editorial review of the paper itself and probably make a pull request with a handful of comments in the next few days and then we'll move to publication.
