### 2025-07-21 aa7b370
include extra references
