+++ paper/paper.bib
@@ -211,6 +211,24 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{conroy&gunn10,
+       author = {{Conroy}, Charlie and {Gunn}, James E.},
+        title = "{The Propagation of Uncertainties in Stellar Population Synthesis Modeling. III. Model Calibration, Comparison, and Evaluation}",
+      journal = {\apj},
+     keywords = {galaxies: evolution, galaxies: stellar content, stars: evolution, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2010,
+        month = apr,
+       volume = {712},
+       number = {2},
+        pages = {833-857},
+          doi = {10.1088/0004-637X/712/2/833},
+archivePrefix = {arXiv},
+       eprint = {0911.3151},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2010ApJ...712..833C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{robotham+20,
        author = {{Robotham}, A.~S.~G. and {Bellstedt}, S. and {Lagos}, C. del P. and {Thorne}, J.~E. and {Davies}, L.~J. and {Driver}, S.~P. and {Bravo}, M.},
         title = "{ProSpect: generating spectral energy distributions with complex star formation and metallicity histories}",
@@ -302,3 +320,169 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{hendriks&Izzard23,
+       author = {{Hendriks}, D. and {Izzard}, R.},
+        title = "{binary\_c-python: A Python-based stellar population synthesis tool and interface to binary\_c}",
+      journal = {The Journal of Open Source Software},
+     keywords = {Python, C, astronomy, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2023,
+        month = may,
+       volume = {8},
+       number = {85},
+          eid = {4642},
+        pages = {4642},
+          doi = {10.21105/joss.04642},
+archivePrefix = {arXiv},
+       eprint = {2306.02779},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023JOSS....8.4642H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@software{ben_johnson_2024_12447779,
+  author       = {{Johnson}, B. and {Foreman-Mackey}, D. and {Sick}, J. and {Leja}, J. and {Walmsley}, M. and {Tollerud}, E. and {Leung}, H. and {Scott}, S. and {Park}, M.},
+  title        = {dfm/python-fsps: v0.4.7},
+  month        = jun,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {v0.4.7},
+  doi          = {10.5281/zenodo.12447779},
+  url          = {https://doi.org/10.5281/zenodo.12447779},
+}
+
+@Article{         harris2020array,
+ title         = {Array programming with {NumPy}},
[…]
