### 2025-07-21 4b31dae
add extra referecences
