+++ paper/paper.bib
@@ -339,7 +339,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@software{ben_johnson_2024_12447779,
+@misc{ben_johnson_2024_12447779,
   author       = {{Johnson}, B. and {Foreman-Mackey}, D. and {Sick}, J. and {Leja}, J. and {Walmsley}, M. and {Tollerud}, E. and {Leung}, H. and {Scott}, S. and {Park}, M.},
   title        = {dfm/python-fsps: v0.4.7},
   month        = jun,
