+++ paper/paper.bib
@@ -341,7 +341,7 @@ archivePrefix = {arXiv},
 
 
 @software{ben_johnson_2024_12447779,
-  author = {Johnson, B. and Foreman-Mackey, D. and Sick, J. and Leja, J. and Walmsley, M. and Tollerud, E. and Leung, H. and Scott, S. and Park, M.},
+  author = {{Johnson}, Benjamin D. and {Foreman-Mackey}, D. and {Sick}, J. and {Leja}, J. and {Walmsley}, M. and {Tollerud}, E. and {Leung}, H. and {Scott}, S. and {Park}, M.},
   title        = {dfm/python-fsps: v0.4.7},
   month        = jun,
   year         = 2024,
