+++ paper/paper.bib
@@ -340,7 +340,7 @@ archivePrefix = {arXiv},
 }
 
 
-@software{ben_johnson_2024_12447779,
+@software{jonhson+24,
   author = {{Johnson}, Benjamin D. and {Foreman-Mackey}, D. and {Sick}, J. and {Leja}, J. and {Walmsley}, M. and {Tollerud}, E. and {Leung}, H. and {Scott}, S. and {Park}, M.},
   title        = {dfm/python-fsps: v0.4.7},
   month        = jun,
