### 2025-07-21 8bc4402
fix multiple citations
