+++ paper/paper.md
@@ -44,7 +44,7 @@ PST is a Python library that offers a comprehensive and flexible framework for s
 
 # State of the field
 
-A number of software packages have been developed to support stellar population synthesis and modeling of galaxy spectral energy distributions. Tools such as [binary_c-python](https://gitlab.surrey.ac.uk/ri0005/binary_c-python) [@hendriks&Izzard23] and [SPISEA](https://github.com/astropy/SPISEA) [@hosek+20] are designed primarily for generating and analyzing simple stellar populations, often with a focus on individual stars, binaries, or star clusters. Meanwhile, libraries such as [python-FSPS](https://github.com/dfm/python-fsps) [@ben_johnson_2024_12447779], a Python interface to the Flexible Stellar Population Synthesis (FSPS) code [@conroy+09, @conroy&gunn10], and the more recent [DSPS](https://github.com/ArgonneCPAC/dsps) [@hearin+23], implemented using JAX for efficient gradient computation and forward modeling, provide extensive modeling capabilities, although they are sometimes limited to a specific set of SSP models or isochrones.
+A number of software packages have been developed to support stellar population synthesis and modeling of galaxy spectral energy distributions. Tools such as [binary_c-python](https://gitlab.surrey.ac.uk/ri0005/binary_c-python) [@hendriks&Izzard23] and [SPISEA](https://github.com/astropy/SPISEA) [@hosek+20] are designed primarily for generating and analyzing simple stellar populations, often with a focus on individual stars, binaries, or star clusters. Meanwhile, libraries such as [python-FSPS](https://github.com/dfm/python-fsps) [@jonhson+24], a Python interface to the Flexible Stellar Population Synthesis (FSPS) code [@conroy+09; @conroy&gunn10], and the more recent [DSPS](https://github.com/ArgonneCPAC/dsps) [@hearin+23], implemented using JAX for efficient gradient computation and forward modeling, provide extensive modeling capabilities, although they are sometimes limited to a specific set of SSP models or isochrones.
 
 Other packages put a stronger emphasis on fitting observed data to derive galaxy properties. Examples of these include  Bayesian frameworks such as [CIGALE](https://cigale.lam.fr/) [@boquien+19], [ProSpect](https://github.com/asgr/ProSpect) [@robotham+20] and [Prospector](https://prospect.readthedocs.io/en/v1.0.0/) [@johnson+21], which infer star formation histories and other physical parameters using spectro-photometric data. Alternative frequentist tools such as [PpXF](https://pypi.org/project/ppxf/) [@capellari+04], [Starlight](http://www.starlight.ufsc.br/) [@cid-fernandes+05], or [Pipe3D](https://gitlab.com/pipe3d/pyPipe3D) [@sanchez+16], are commonly used to extract stellar kinematics and stellar population parameters from observed galaxy spectra, often in the context of integral field spectroscopy.
 
@@ -95,6 +95,6 @@ We acknowledge financial support from the Spanish State Research Agency (AEI/10.
 
 Daniel Jiménez-López was supported by Fondo Europeo de Desarrollo Regional (MINCIN/AEI/10.13039/501100011033/FEDER, UE), through a FPI-contract fellowship in the project PID2022-138560NB.
 
-Our package relies on several widely used open-source Python libraries, including [Numpy](http://www.numpy.org) [@harris2020array], [Matplotlib](https://www.matplotlib.org/) [@hunter:2007] and [Astropy](http://www.astropy.org) [@astropy:2013, @astropy:2018, @astropy:2022].
[…]
