### 2025-07-21 #8203 @PabloCorcho
@ivastar the PR has been merged, and the remaining editorial comments have been implemented in https://github.com/paranoya/population-synthesis-toolkit/pull/75
