### 2025-07-23 #8203 @ivastar
@PabloCorcho a couple last notes about the references:
- The Filter Profile Service mentions 3 references, but only one is included in the paper. Can you please add the other 2?
- I don't think that minting a DOI for PyKOALA now prevents you from a future JOSS submission and would allow you to have a citable entity in the mean time, but maybe @warrickball can comment?
