### 2025-07-23 6011347
add extra SVO references
