+++ paper/paper.bib
@@ -70,6 +70,20 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+
+@MISC{rodrigo+12,
+       author = {{Rodrigo}, Carlos and {Solano}, Enrique and {Bayo}, Amelia},
+        title = "{SVO Filter Profile Service Version 1.0}",
+ howpublished = {IVOA Working Draft 15 October 2012},
+         year = 2012,
+        month = oct,
+        pages = {1015},
+          doi = {10.5479/ADS/bib/2012ivoa.rept.1015R},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2012ivoa.rept.1015R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
 @INPROCEEDINGS{rodrigo+20,
        author = {{Rodrigo}, C. and {Solano}, E.},
         title = "{The SVO Filter Profile Service}",
@@ -79,6 +93,24 @@ archivePrefix = {arXiv},
           eid = {182},
         pages = {182},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020sea..confE.182R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{rodrigo+24,
+       author = {{Rodrigo}, Carlos and {Cruz}, Patricia and {Aguilar}, John F. and {Aller}, Alba and {Solano}, Enrique and {G{\'a}lvez-Ortiz}, Maria Cruz and {Jim{\'e}nez-Esteban}, Francisco and {Mas-Buitrago}, Pedro and {Bayo}, Amelia and {Cort{\'e}s-Contreras}, Miriam and {Murillo-Ojeda}, Raquel and {Bonoli}, Silvia and {Cenarro}, Javier and {Dupke}, Renato and {L{\'o}pez-Sanjuan}, Carlos and {Mar{\'\i}n-Franch}, Antonio and {de Oliveira}, Claudia Mendes and {Moles}, Mariano and {Taylor}, Keith and {Varela}, Jes{\'u}s and {Rami{\'o}}, H{\'e}ctor V{\'a}zquez},
+        title = "{Photometric segregation of dwarf and giant FGK stars using the SVO Filter Profile Service and photometric tools}",
+      journal = {\aap},
+     keywords = {methods: data analysis, techniques: photometric, astronomical databases: miscellaneous, virtual observatory tools, stars: fundamental parameters, stars: late-type, Solar and Stellar Astrophysics, Astrophysics of Galaxies},
+         year = 2024,
+        month = sep,
+       volume = {689},
+          eid = {A93},
+        pages = {A93},
+          doi = {10.1051/0004-6361/202449998},
+archivePrefix = {arXiv},
+       eprint = {2406.03310},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024A&A...689A..93R},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
