+++ paper/paper.bib
@@ -72,7 +72,7 @@ archivePrefix = {arXiv},
 
 
 @MISC{rodrigo+12,
-       author = {{Rodrigo}, Carlos and {Solano}, Enrique and {Bayo}, Amelia},
+       author = {{Rodrigo}, C. and {Solano}, E. and {Bayo}, A.},
         title = "{SVO Filter Profile Service Version 1.0}",
  howpublished = {IVOA Working Draft 15 October 2012},
          year = 2012,
@@ -97,7 +97,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{rodrigo+24,
-       author = {{Rodrigo}, Carlos and {Cruz}, Patricia and {Aguilar}, John F. and {Aller}, Alba and {Solano}, Enrique and {G{\'a}lvez-Ortiz}, Maria Cruz and {Jim{\'e}nez-Esteban}, Francisco and {Mas-Buitrago}, Pedro and {Bayo}, Amelia and {Cort{\'e}s-Contreras}, Miriam and {Murillo-Ojeda}, Raquel and {Bonoli}, Silvia and {Cenarro}, Javier and {Dupke}, Renato and {L{\'o}pez-Sanjuan}, Carlos and {Mar{\'\i}n-Franch}, Antonio and {de Oliveira}, Claudia Mendes and {Moles}, Mariano and {Taylor}, Keith and {Varela}, Jes{\'u}s and {Rami{\'o}}, H{\'e}ctor V{\'a}zquez},
+       author = {{Rodrigo}, C. and {Cruz}, P. and {Aguilar}, J. F. and {Aller}, A. and {Solano}, E. and {G{\'a}lvez-Ortiz}, M. C. and {Jim{\'e}nez-Esteban}, F. and {Mas-Buitrago}, P. and {Bayo}, A. and {Cort{\'e}s-Contreras}, M. and {Murillo-Ojeda}, R. and {Bonoli}, S. and {Cenarro}, J. and {Dupke}, R. and {L{\'o}pez-Sanjuan}, C. and {Mar{\'\i}n-Franch}, A. and {de Oliveira}, C. M. and {Moles}, M. and {Taylor}, K. and {Varela}, J. and {Rami{\'o}}, H. V.,
         title = "{Photometric segregation of dwarf and giant FGK stars using the SVO Filter Profile Service and photometric tools}",
       journal = {\aap},
      keywords = {methods: data analysis, techniques: photometric, astronomical databases: miscellaneous, virtual observatory tools, stars: fundamental parameters, stars: late-type, Solar and Stellar Astrophysics, Astrophysics of Galaxies},
