+++ paper/paper.bib
@@ -97,7 +97,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{rodrigo+24,
-       author = {{Rodrigo}, C. and {Cruz}, P. and {Aguilar}, J. F. and {Aller}, A. and {Solano}, E. and {G{\'a}lvez-Ortiz}, M. C. and {Jim{\'e}nez-Esteban}, F. and {Mas-Buitrago}, P. and {Bayo}, A. and {Cort{\'e}s-Contreras}, M. and {Murillo-Ojeda}, R. and {Bonoli}, S. and {Cenarro}, J. and {Dupke}, R. and {L{\'o}pez-Sanjuan}, C. and {Mar{\'\i}n-Franch}, A. and {de Oliveira}, C. M. and {Moles}, M. and {Taylor}, K. and {Varela}, J. and {Rami{\'o}}, H. V.,
+       author = {{Rodrigo}, C. and {Cruz}, P. and {Aguilar}, J. F. and {Aller}, A. and {Solano}, E. and {G{\'a}lvez-Ortiz}, M. C. and {Jim{\'e}nez-Esteban}, F. and {Mas-Buitrago}, P. and {Bayo}, A. and {Cort{\'e}s-Contreras}, M. and {Murillo-Ojeda}, R. and {Bonoli}, S. and {Cenarro}, J. and {Dupke}, R. and {L{\'o}pez-Sanjuan}, C. and {Mar{\'\i}n-Franch}, A. and {de Oliveira}, C. M. and {Moles}, M. and {Taylor}, K. and {Varela}, J. and {Rami{\'o}}, H. V.},
         title = "{Photometric segregation of dwarf and giant FGK stars using the SVO Filter Profile Service and photometric tools}",
       journal = {\aap},
      keywords = {methods: data analysis, techniques: photometric, astronomical databases: miscellaneous, virtual observatory tools, stars: fundamental parameters, stars: late-type, Solar and Stellar Astrophysics, Astrophysics of Galaxies},
