+++ paper/paper.bib
@@ -198,15 +198,6 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@misc{pykoala,
-  author = {{Corcho-Caballero}, P. and {Ascasibar}, Y. and {L\'opez-S\'anchez}, A. R. and {Gonz\'alez-Bol\'ivar}, M., and {Lorente}, N., and {Tocknell}, J. and {Jim\'enez-Ibarra}, F. and {Daluwathumullagamage}, P. J. and {Quattropani}, G. and {Owers}, M},
-  title = {PyKOALA},
-  year = {2025},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/pykoala}
-}
-
 @ARTICLE{hearin+23,
        author = {{Hearin}, Andrew P. and {Chaves-Montero}, Jon{\'a}s and {Alarcon}, Alex and {Becker}, Matthew R. and {Benson}, Andrew},
         title = "{DSPS: Differentiable stellar population synthesis}",
@@ -519,3 +510,24 @@ archivePrefix = {arXiv},
   doi       = {10.1109/MCSE.2007.55},
   year      = 2007
 }
+
+@misc{pykoala_adass,
+      title={The PyKOALA python library: a multi-instrument package for IFS data reduction}, 
+      author={{Corcho-Caballero}, P. and {Ascasibar}, Y. and {L\'opez-S\'anchez}, A. R. and {Gonz\'alez-Bol\'ivar}, M., and {Lorente}, N., and {Tocknell}, J. and {Jim\'enez-Ibarra}, F. and {Daluwathumullagamage}, P. J. and {Quattropani}, G. and {Owers}, M and {Verdoes-Kleijn}, G. A.},
+      year={2025},
+      eprint={2507.18347},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.IM},
+      url={https://arxiv.org/abs/2507.18347}, 
+}
+
+@misc{pykoala,
+  author = {{Corcho-Caballero}, P. and {Ascasibar}, Y. and {L\'opez-S\'anchez}, A. R. and {Gonz\'alez-Bol\'ivar}, M., and {Lorente}, N., and {Tocknell}, J. and {Jim\'enez-Ibarra}, F. and {Daluwathumullagamage}, P. J. and {Quattropani}, G. and {Owers}, M},
+  title = {PyKOALA},
+  year = {2025},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/pykoala}
+}
+
+
