### 2025-07-27 2a036ff
Initial editorial tweaks for JOSS publication
