+++ paper/paper.md
@@ -7,12 +7,12 @@ tags:
   - galaxies
   - stellar population synthesis
 authors:
-  - name: Pablo Corcho-Caballero 
+  - name: Pablo Corcho-Caballero
     orcid: 0000-0001-6327-7080
     corresponding: true
     equal-contrib: true
     affiliation: 1 # (Multiple affiliations must be quoted)
-  - name: Yago Ascasibar 
+  - name: Yago Ascasibar
     orcid: 0000-0003-1577-2479
     corresponding: true
     equal-contrib: true
@@ -21,9 +21,9 @@ authors:
     orcid: 0009-0001-2907-8691
     corresponding: true
     equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
-    affiliation: 4 
+    affiliation: 4
 affiliations:
- - name: Kapteyn Astronomical Institute, University of Groningen, the Netherlands
+ - name: Kapteyn Astronomical Institute, University of Groningen, The Netherlands
    index: 1
  - name: Department of Theoretical Physics, Universidad Autónoma de Madrid (UAM), Spain
    index: 2
@@ -46,7 +46,7 @@ PST is a Python library that offers a comprehensive and flexible framework for s
 
 A number of software packages have been developed to support stellar population synthesis and modeling of galaxy spectral energy distributions. Tools such as [binary_c-python](https://gitlab.surrey.ac.uk/ri0005/binary_c-python) [@hendriks&Izzard23] and [SPISEA](https://github.com/astropy/SPISEA) [@hosek+20] are designed primarily for generating and analyzing simple stellar populations, often with a focus on individual stars, binaries, or star clusters. Meanwhile, libraries such as [python-FSPS](https://github.com/dfm/python-fsps) [@jonhson+24], a Python interface to the Flexible Stellar Population Synthesis (FSPS) code [@conroy+09; @conroy&gunn10], and the more recent [DSPS](https://github.com/ArgonneCPAC/dsps) [@hearin+23], implemented using JAX for efficient gradient computation and forward modeling, provide extensive modeling capabilities, although they are sometimes limited to a specific set of SSP models or isochrones.
 
-Other packages put a stronger emphasis on fitting observed data to derive galaxy properties. Examples of these include  Bayesian frameworks such as [CIGALE](https://cigale.lam.fr/) [@boquien+19], [ProSpect](https://github.com/asgr/ProSpect) [@robotham+20] and [Prospector](https://prospect.readthedocs.io/en/v1.0.0/) [@johnson+21], which infer star formation histories and other physical parameters using spectro-photometric data. Alternative frequentist tools such as [PpXF](https://pypi.org/project/ppxf/) [@capellari+04], [Starlight](http://www.starlight.ufsc.br/) [@cid-fernandes+05], or [Pipe3D](https://gitlab.com/pipe3d/pyPipe3D) [@sanchez+16], are commonly used to extract stellar kinematics and stellar population parameters from observed galaxy spectra, often in the context of integral field spectroscopy.
[…]
